Split gallery surface components
